backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
Pillow>=10.2.0
aiofiles==23.2.1
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
